requirements.txt
tweepy==4.14.0
python-dotenv==1.0.1
APScheduler==3.10.4
google-genai>=1.0.0
feedparser>=6.0.0
